requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# OAuth
authlib==1.3.0
httpx==0.26.0
google-auth==2.45.0

# WebSocket & Real-time
python-socketio==5.11.0
python-engineio==4.9.0

# Google Gemini AI
google-generativeai==0.3.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
aiofiles==23.2.1
pypdf==3.17.4

# Development & Code Quality
pyright==1.1.348
black==24.1.1
isort==5.13.2
flake8==7.0.0
pre-commit==3.6.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring & Logging
structlog==24.1.0
